chore: added logic to fire event if a percentage of the current budget is exceeded when creating an expense, added kafka logic for producing and consuming event

+++ pocketpipo/src/main/java/com/pocketpipo/pocketpipo/dto/CreateBudgetRequestDTO.java
@@ -1,15 +1,16 @@
 package com.pocketpipo.pocketpipo.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class CreateBudgetRequestDTO {
     private String name;
-    private long maxAmount;
+    private BigDecimal maxAmount;
     private long userId;
     private LocalDate startDate;
     private LocalDate endDate;
 
-    public CreateBudgetRequestDTO(String name, long maxAmount, long userId, LocalDate startDate, LocalDate endDate) {
+    public CreateBudgetRequestDTO(String name, BigDecimal maxAmount, long userId, LocalDate startDate, LocalDate endDate) {
         this.name = name;
         this.maxAmount = maxAmount;
         this.userId = userId;
@@ -23,10 +24,10 @@ public class CreateBudgetRequestDTO {
     public void setName(String name) {
         this.name = name;
     }
-    public long getMaxAmount() {
+    public BigDecimal getMaxAmount() {
         return maxAmount;
     }
-    public void setMaxAmount(long maxAmount) {
+    public void setMaxAmount(BigDecimal maxAmount) {
         this.maxAmount = maxAmount;
     }
     public long getUserId() {
